Initial Commit
Seperate code for training and deploying a network that learns about
TORCS from the Caffe code. Now we have an up to date definition of the
network used for training and of the solver. Instead of saving image
_and_ regression targets in a single Datum, we use two leveldb databases
where one contains Datums with the image data and the other contains
Datums with the regression targets. The goal is to run the whole
DeepDriving Project without modifications to the Caffe code but use it
as a library as it is. The visualization code is very minimal, split is
used to split a dataset that was generated from the old codebase and
contains image and regression data in a single Datum. The normalize code
can be run on the resulting regression target database to normalize the
targets to a specified interval.